train_gcp/requirements.txt
pydicom
keras
google-cloud-storage
numpy
setuptools
pandas
argparse
Pillow
sklearn